src/auth/requirements.txt
black==23.1.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
Deprecated==1.2.13
Flask==2.2.3
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
tomli==2.0.1
Werkzeug==2.2.3
wrapt==1.15.0
